tegra-storage-layout: fix support for large rootfs
When using a rootfs with external device too large for the emmc
and a redundant rootfs partition, the APP_b partition is not
removed from the emmc partition and therefore bup payload
creation fails when it tries to place the too large rootfs into
the emmc.

Expand the existing logic to remove the APP_b partition, if it
exists, in addition to the APP partition.

Closes https://github.com/OE4T/meta-tegra/issues/1439

Signed-off-by: Dan Walkes <danwalkes@trellis-logic.com>